fixed problem 2 in c++

+++ lib/c/factor.h
@@ -4,7 +4,7 @@
 #include <stdbool.h>
 
 bool even(int n);
-bool even_l(int n);
+bool even_l(long n);
 
 bool odd(int n);
 bool odd_l(long n);
